Revert "add : sort troop setting view GroupAppInfoListView(GroupFileView) to top"
This reverts commit 953419e5
