+++ app/src/main/java/top/linl/hook/OffRelationshipIdentification.java
@@ -11,8 +11,8 @@ import io.github.qauxv.dsl.FunctionEntryRouter;
 import io.github.qauxv.hook.CommonSwitchFunctionHook;
 import io.github.qauxv.util.Initiator;
 import java.lang.reflect.Method;
-import top.linl.util.reflect.FieIdUtils;
-import top.linl.util.reflect.MethodTool;
+import top.linl.util.FieIdUtils;
+import top.linl.util.MethodTool;
 
 
 @FunctionHookEntry
