--- app/src/main/java/top/linl/util/ScreenParamUtils.java
@@ -1,117 +1,0 @@
-package top.linl.util;
-
-import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.content.Context;
-import android.content.res.Resources;
-import android.graphics.Color;
-import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
-
-public class ScreenParamUtils {
-
-
-    /**
-     * sp转px
-     */
-    public static int spToPx(Context context, float spValue) {
-        final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (spValue * fontScale + 0.5f);
-    }
-
-    /**
-     * 获取手机像素宽度
-     *
-     * @return 宽度px
-     */
-    public static int getScreenWidth(Context context) {
-        return context.getResources().getDisplayMetrics().widthPixels;
-    }
-
-    /**
-     * 获取手机像素高度(不包含虚拟导航栏和状态栏)
-     *
-     * @return 高度px
-     */
-    public static int getScreenHeight(Context context) {
-        return context.getResources().getDisplayMetrics().heightPixels;
-    }
-
-
-    /**
-     * 获取导航栏高度
-     *
-     * @return px
-     */
-    public static int getNavigationBarHeight(Context context) {
-        int navigationBarHeight = -1;
-        Resources resources = context.getResources();
-        @SuppressLint({"DiscouragedApi", "InternalInsetResource"}) int resourceId = resources.getIdentifier("navigation_bar_height", "dimen", "android");
-        if (resourceId > 0) {
-            navigationBarHeight = resources.getDimensionPixelSize(resourceId);
-        }
-        return navigationBarHeight;
-    }
-
-    /**
-     * 获取状态栏高度
-     *
-     * @return px
-     */
-    public static int getStatusBarHeight(Context context) {
-        int height = 0;
-        @SuppressLint({"InternalInsetResource", "DiscouragedApi"})
-        int resourceId = context.getResources().getIdentifier("status_bar_height", "dimen", "android");
-        if (resourceId > 0) {
-            height = context.getResources().getDimensionPixelSize(resourceId);
-        }
-        return height;
-    }
-
-    /**
-     * 根据手机的分辨率从 dp(相对大小) 的单位 转成为 px(像素)
-     */
-    public static int dpToPx(Context context, float dpValue) {
-        // 获取屏幕密度
-        final float scale = context.getResources().getDisplayMetrics().density;
-        // 结果+0.5是为了int取整时更接近
-        return (int) (dpValue * scale + 0.5f);
-    }
-
-
-    /**
-     * 根据手机的分辨率从 px(像素) 的单位 转成为 dp(相对大小)
-     */
-    public static int pxToDp(Context context, float pxValue) {
-        final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (pxValue / scale + 0.5f);
-    }
-
-    /**
-     * px转sp
-     */
-    public static int pxToSp(Context context, float pxValue) {
-        final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (pxValue / fontScale + 0.5f);
-    }
-
-    /**
-     * 将当前活动设置为透明的状态栏
-     *
-     * @param activity 当前Activity
-     */
-    protected void requestTranslucentStatusBar(Activity activity) {
-        Window window = activity.getWindow();
-        window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS
-                | WindowManager.LayoutParams.FLAG_TRANSLUCENT_NAVIGATION);
-        window.getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                | View.SYSTEM_UI_FLAG_LAYOUT_STABLE);
-        window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-        window.setStatusBarColor(Color.TRANSPARENT);
-        window.setNavigationBarColor(Color.TRANSPARENT);
-    }
-
-
-}
